requirements.txt
pandas
numpy
torch
torchvision
transformers
stockstats
PyYAML
ast  # part of Python standard library, but note it here if you use it extensively